[AIRFLOW-5787] Moving AWS SQS to /providers/aws.
