+++ airflow/contrib/sensors/aws_sqs_sensor.py
@@ -17,77 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from airflow.contrib.hooks.aws_sqs_hook import SQSHook
-from airflow.exceptions import AirflowException
-from airflow.sensors.base_sensor_operator import BaseSensorOperator
-from airflow.utils.decorators import apply_defaults
+"""This module is deprecated. Please use `airflow.providers.aws.sensors.sqs`."""
 
+import warnings
 
-class SQSSensor(BaseSensorOperator):
-    """
-    Get messages from an SQS queue and then deletes  the message from the SQS queue.
-    If deletion of messages fails an AirflowException is thrown otherwise, the message
-    is pushed through XCom with the key ``message``.
+# pylint: disable=unused-import
+from airflow.providers.aws.sensors.sqs import SQSSensor  # noqa
 
-    :param aws_conn_id: AWS connection id
-    :type aws_conn_id: str
-    :param sqs_queue: The SQS queue url (templated)
-    :type sqs_queue: str
-    :param max_messages: The maximum number of messages to retrieve for each poke (templated)
-    :type max_messages: int
-    :param wait_time_seconds: The time in seconds to wait for receiving messages (default: 1 second)
-    :type wait_time_seconds: int
-    """
-
-    template_fields = ('sqs_queue', 'max_messages')
-
-    @apply_defaults
-    def __init__(self,
-                 sqs_queue,
-                 aws_conn_id='aws_default',
-                 max_messages=5,
-                 wait_time_seconds=1,
-                 *args,
-                 **kwargs):
-        super().__init__(*args, **kwargs)
-        self.sqs_queue = sqs_queue
-        self.aws_conn_id = aws_conn_id
-        self.max_messages = max_messages
-        self.wait_time_seconds = wait_time_seconds
-
-    def poke(self, context):
-        """
-        Check for message on subscribed queue and write to xcom the message with key ``messages``
-
-        :param context: the context object
-        :type context: dict
-        :return: ``True`` if message is available or ``False``
-        """
-
-        sqs_hook = SQSHook(aws_conn_id=self.aws_conn_id)
-        sqs_conn = sqs_hook.get_conn()
-
-        self.log.info('SQSSensor checking for message on queue: %s', self.sqs_queue)
-
-        messages = sqs_conn.receive_message(QueueUrl=self.sqs_queue,
-                                            MaxNumberOfMessages=self.max_messages,
-                                            WaitTimeSeconds=self.wait_time_seconds)
-
-        self.log.info("received message %s", str(messages))
-
-        if 'Messages' in messages and len(messages['Messages']) > 0:
-
-            entries = [{'Id': message['MessageId'], 'ReceiptHandle': message['ReceiptHandle']}
-                       for message in messages['Messages']]
-
-            result = sqs_conn.delete_message_batch(QueueUrl=self.sqs_queue,
-                                                   Entries=entries)
-
-            if 'Successful' in result:
-                context['ti'].xcom_push(key='messages', value=messages)
-                return True
-            else:
-                raise AirflowException(
-                    'Delete SQS Messages failed ' + str(result) + ' for messages ' + str(messages))
-
-        return False
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.aws.sensors.sqs`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
