+++ docs/operators-and-hooks-ref.rst
@@ -388,9 +388,9 @@ These integrations allow you to perform various operations within the Amazon Web
      -
 
    * - `Amazon Simple Queue Service (SQS) <https://aws.amazon.com/sns/>`__
-     - :mod:`airflow.contrib.hooks.aws_sqs_hook`
-     - :mod:`airflow.contrib.operators.aws_sqs_publish_operator`
-     - :mod:`airflow.contrib.sensors.aws_sqs_sensor`
+     - :mod:`airflow.providers.aws.hooks.sqs`
+     - :mod:`airflow.providers.aws.operators.sqs`
+     - :mod:`airflow.providers.aws.sensors.sqs`
 
 Transfer operators and hooks
 ''''''''''''''''''''''''''''
