+++ tests/test_core_to_contrib.py
@@ -136,6 +136,10 @@ HOOK = [
         "airflow.providers.aws.hooks.s3.S3Hook",
         "airflow.hooks.S3_hook.S3Hook",
     ),
+    (
+        "airflow.providers.aws.hooks.sqs.SQSHook",
+        "airflow.contrib.hooks.aws_sqs_hook.SQSHook",
+    ),
 ]
 OPERATOR = [
     (
@@ -746,6 +750,10 @@ OPERATOR = [
         "airflow.providers.aws.operators.athena.AWSAthenaOperator",
         "airflow.contrib.operators.aws_athena_operator.AWSAthenaOperator",
     ),
+    (
+        "airflow.providers.aws.operators.sqs.SQSPublishOperator",
+        "airflow.contrib.operators.aws_sqs_publish_operator.SQSPublishOperator",
+    )
 ]
 SENSOR = [
     (
@@ -787,6 +795,11 @@ SENSOR = [
         "airflow.providers.aws.sensors.athena.AthenaSensor",
         "airflow.contrib.sensors.aws_athena_sensor.AthenaSensor",
     ),
+    (
+        "airflow.providers.aws.sensors.sqs.SQSSensor",
+        "airflow.contrib.sensors.aws_sqs_sensor.SQSSensor",
+    ),
+
 ]
 ALL = HOOK + OPERATOR + SENSOR
 RENAMED_HOOKS = [
